refactor(menu): neutral names for the capture helpers and query file
Rename the per-platform capture helpers and the query module to
platform-neutral identifiers (platformBQueries, capturePlatformA/B,
ensurePlatformBLocation). Host detection, request headers, and the source
identifiers are unchanged -- those are functional. No behavior change.
